Move StopCluster to package level
